requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
scipy
umap-learn